tophat: fixes for python and script paths
This adds patches to the tophat recipe to make all the python scripts
look for "python2.7" instead of just "python", since otherwise the
tophat scripts only run under python 2, and will fail if "python" points
to python 3. In addition, the "tophat2" script that just acts as an
alias to "tophat" has been modified to use `readlink -f` instead of just
`readlink`, so it now works when symlinked.

This also adds the missing dependency on python.